my_first_ai_app/requirements.txt
flask
openai
google-generativeai
python-dotenv
